move shared code to a common place
