+++ Sources/App/Controllers/ClientController.swift
@@ -140,77 +140,6 @@ struct ClientController: APIRouteCollection {
 		return promise.futureResult
 	}
 
-    /// Prometheus webhook alert object. Applied from https://prometheus.io/docs/alerting/latest/configuration/#webhook_config
-    public struct AlertmanagerAlert: Content {
-        var status: String
-        var labels: [String:String]
-        var annotations: [String:String]
-        var startsAt: String
-        var endsAt: String
-        // Identifies the entity that caused the alert.
-        var generatorURL: String
-        // Fingerprint to identify the alert.
-        var fingerprint: String
-    }
-
-    /// Prometheus Amertmanager webhook payload. Applied from https://prometheus.io/docs/alerting/latest/configuration/#webhook_config
-    public struct AlertmanagerWebhookPayload: Content {
-        var version: String
-        /// Key identifying the group of alerts (e.g. to deduplicate).
-        var groupKey: String
-        /// How many alerts have been truncated due to "max_alerts".
-        var truncatedAlerts: Int
-        var status: String
-        var receiver: String
-        var groupLabels: [String:String]
-        var commonLabels: [String:String]
-        var commonAnnotations: [String:String]
-        /// backlink to the Alertmanager.
-        var externalURL: String
-        var alerts: [AlertmanagerAlert]
-    }
-
-    func buildFezData(from fez: FriendlyFez, with pivot: FezParticipant? = nil, posts: [FezPostData]? = nil, 
-            for cacheUser: UserCacheData, on req: Request) throws -> FezData {
-        let userBlocks = cacheUser.getBlocks()
-        // init return struct
-        let ownerHeader = try req.userCache.getHeader(fez.$owner.id)
-        var fezData : FezData = try FezData(fez: fez, owner: ownerHeader)
-        if pivot != nil || (cacheUser.accessLevel.hasAccess(.moderator) && fez.fezType != .closed) {
-            let allParticipantHeaders = req.userCache.getHeaders(fez.participantArray)
-
-            // masquerade blocked users
-            let valids = allParticipantHeaders.map { (member: UserHeader) -> UserHeader in
-                if userBlocks.contains(member.userID) {
-                    return UserHeader.Blocked
-                }
-                return member
-            }
-            // populate fezData's participant list and waiting list
-            var participants: [UserHeader]
-            var waitingList: [UserHeader]
-            if valids.count > fez.maxCapacity && fez.maxCapacity > 0 {
-                participants = Array(valids[valids.startIndex..<fez.maxCapacity])
-                waitingList = Array(valids[fez.maxCapacity..<valids.endIndex])
-            }
-            else {
-                participants = valids
-                waitingList = []
-            }
-            fezData.members = FezData.MembersOnlyData(participants: participants, waitingList: waitingList, 
-                    postCount: fez.postCount - (pivot?.hiddenCount ?? 0), readCount: pivot?.readCount ?? 0, posts: posts)
-        }
-       
-       
-        return fezData
-    }
-
-    // func getUserPivot(fez: FriendlyFez, userID: UUID, on db: Database) -> EventLoopFuture<FezParticipant?> {
-	// 	return fez.$participants.$pivots.query(on: db)
-	// 			.filter(\.$user.$id == userID)
-	// 			.first()
-	// }
-
     func prometheusAlertHandler(_ req: Request) async throws -> Response {
         // let futureString: EventLoopFuture<String> = "Hello"
         // return EventLoopFuture<String>("hello")
@@ -243,33 +172,7 @@ struct ClientController: APIRouteCollection {
         try await post.save(on: req.db)
         // try await fez.save(on: req.db)
 
-        
-
-        // try await fez.save(on: req.db).flatMap { _ in
-		// 	return User.query(on: req.db).filter(\.$id ~~ initialUsers).all().flatMap { participants in
-		// 		return fez.$participants.attach(participants, on: req.db, { $0.readCount = 0; $0.hiddenCount = 0 }).throwingFlatMap { (_) in
-		// 			return fez.$participants.$pivots.query(on: req.db).filter(\.$user.$id == sourceUser.userID)
-		// 					.first().flatMapThrowing() { creatorPivot -> FezParticipant in
-		// 				let fezData = try buildFezData(from: fez, with: creatorPivot, posts: [], for: req.userCache.getUser(username: "client")!, on: req)
-		// 				// with 201 status
-		// 				let response = Response(status: .created)
-		// 				try response.content.encode(fezData)
-		// 			}
-		// 		}
-		// 	}
-		// }
-
-        // try await getUserPivot(fez: fez, userID: sourceUser.userID, on: req.db).flatMapThrowing { pivot -> Void in
-        //     // A user posting is assumed to have read all prev posts. (even if this proves untrue, we should increment
-        //     // readCount as they've read the post they just wrote!)
-        //     if let pivot = pivot {
-        //         pivot.readCount = fez.postCount - pivot.hiddenCount
-        //         _ = pivot.save(on: req.db)
-        //         print("doing the thing?")
-        //     }
-        //     // return try FezPostData(post: post, author: effectiveAuthor.makeHeader())
-        // }
-
+        // THIS IS THE GOOD ONE
         try await fez.save(on: req.db).flatMap { _ in
 			return User.query(on: req.db).filter(\.$id ~~ initialUsers).all().flatMap { participants in
 				return fez.$participants.attach(participants, on: req.db, { $0.readCount = 0; $0.hiddenCount = 0 }).throwingFlatMap { (_) in
@@ -289,19 +192,12 @@ struct ClientController: APIRouteCollection {
 
         print("Done?")
 
-        // let fezContent = FezContentData(fezType: .closed, title: formContent.subject, info: "", startTime: nil, endTime: nil,
-				// location: nil, minCapacity: 0, maxCapacity: 0, initialUsers: participants)
-    	// return apiQuery(req, endpoint: "/fez/create", method: .POST, beforeSend: { req throws i
-        // let fez = FriendlyFez(owner: user.userID, fezType: data.fezType, title: data.title, info: data.info,
-				// location: data.location, startTime: data.startTime, endTime: data.endTime,
-				// minCapacity: data.minCapacity, maxCapacity: data.maxCapacity)
-		// This filters out anyone on the creator's blocklist and any duplicate IDs.
-		// var creatorBlocks = user.getBlocks()
-		// let initialUsers = ([user.userID] + data.initialUsers).filter { creatorBlocks.insert($0).inserted }
-		// fez.participantArray = initialUsers
-
         return Response(status: .ok)
     }
 
     // MARK: - Helper Functions
+}
+
+extension ClientController: FezProtocol {
+    /// This page intentionally left blank
 }
